help refactor - kamangir/bolt#746
