+++ blue_geo/help/catalog/__init__.py
@@ -1,0 +1,12 @@
+from blue_geo.help.catalog.browse import help_browse
+from blue_geo.help.catalog.get import help_get
+from blue_geo.help.catalog.list import help_list
+from blue_geo.help.catalog.query import help_query
+
+
+help_functions = {
+    "browse": help_browse,
+    "get": help_get,
+    "list": help_list,
+    "query": help_query,
+}
