+++ setup.py
@@ -33,6 +33,7 @@ setup(
         f"{NAME}.datacube.label",
         #
         f"{NAME}.help",
+        f"{NAME}.help.catalog",
         f"{NAME}.help.datacube",
         f"{NAME}.help.watch",
         f"{NAME}.help.QGIS",
